Added doors to the airport map
