+++ src/mock/data/building-map.ts
@@ -1,6 +1,8 @@
 import * as RomiCore from '@osrf/romi-js-core-interfaces';
 import airportPngUrl from './airport_terminal.png';
 import officePngUrl from './office.png';
+import officeMap from './building-map-office';
+import airportMap from './building-map-airport';
 
 export default async function buildingMap(): Promise<RomiCore.BuildingMap> {
   let officePng: Uint8Array;
@@ -15,120 +17,6 @@ export default async function buildingMap(): Promise<RomiCore.BuildingMap> {
     const airportPngResp = await fetch(airportPngUrl);
     airportPng = new Uint8Array(await airportPngResp.arrayBuffer());
   }
-
-  const doors = [
-    {
-      name: 'main_door',
-      v1_x: 8.2,
-      v1_y: -5.5,
-      v2_x: 7.85,
-      v2_y: -6.2,
-      door_type: 2,
-      motion_range: -1.571,
-      motion_direction: 1,
-    },
-    {
-      name: 'hardware_door',
-      v1_x: 4.9,
-      v1_y: -4,
-      v2_x: 4.4,
-      v2_y: -5,
-      door_type: 1,
-      motion_range: 1.571,
-      motion_direction: -1,
-    },
-    {
-      name: 'coe_door',
-      v1_x: 19.5,
-      v1_y: -10.8,
-      v2_x: 19.5,
-      v2_y: -9.9,
-      door_type: 1,
-      motion_range: 1.571,
-      motion_direction: 1,
-    },
-    {
-      name: 'exit_door',
-      v1_x: 12.2,
-      v1_y: -2.7,
-      v2_x: 14.1,
-      v2_y: -2.7,
-      door_type: 1,
-      motion_range: -1.571,
-      motion_direction: 1,
-    },
-  ];
-
-  const lifts = [
-    {
-      name: 'Lift1',
-      doors: [{
-        name: 'lift1_front_door',
-        v1_x: 8.80,
-        v1_y: -2.3,
-        v2_x: 7.70,
-        v2_y: -4.5,
-        door_type: 1,
-        motion_range: 0,
-        motion_direction: 1,
-      }],
-      levels: ['L1', 'L2', 'L3'],
-      ref_x: 7.1,
-      ref_y: -2.8,
-      ref_yaw: 0.5,
-      width: 2.5,
-      depth: 2.5,
-      wall_graph: {
-        name: 'wallgraph',
-        vertices: [],
-        edges: [],
-        params: [],
-      },
-    },
-    {
-      name: 'Lift2',
-      doors: [
-        {
-          name: 'lift2_front_door',
-          v1_x: 8.95,
-          v1_y: -12.38,
-          v2_x: 10.08,
-          v2_y: -12.38,
-          door_type: 1,
-          motion_range: 0,
-          motion_direction: 1,
-        }
-      ],
-      levels: ['L1', 'L2', 'L3', 'L4'],
-      ref_x: 9.5,
-      ref_y: -13,
-      ref_yaw: 1.571,
-      width: 1,
-      depth: 1,
-      wall_graph: {
-        name: 'wallgraph',
-        vertices: [],
-        edges: [],
-        params: [],
-      },
-    },
-    {
-      name: 'mysterious_lift',
-      doors: [],
-      levels: ['L1', 'L2', 'L3', 'L4'],
-      ref_x: 22,
-      ref_y: -10,
-      ref_yaw: 1.571,
-      width: 1,
-      depth: 1,
-      wall_graph: {
-        name: 'wallgraph',
-        vertices: [],
-        edges: [],
-        params: [],
-      },
-    },
-  ];
 
   return {
     name: 'test building',
@@ -147,25 +35,8 @@ export default async function buildingMap(): Promise<RomiCore.BuildingMap> {
             yaw: 0,
           },
         ],
-        places: [
-          {
-            name: 'Place1',
-            x: 2,
-            y: -2,
-            yaw: 0,
-            position_tolerance: 0,
-            yaw_tolerance: 0,
-          },
-          {
-            name: 'Place2',
-            x: 8,
-            y: -4,
-            yaw: 0,
-            position_tolerance: 0,
-            yaw_tolerance: 0,
-          },
-        ],
-        doors: doors,
+        doors: officeMap.doors,
+        places: officeMap.places,
         nav_graphs: [],
         wall_graph: {
           name: 'wallgraph',
@@ -177,7 +48,6 @@ export default async function buildingMap(): Promise<RomiCore.BuildingMap> {
       {
         name: 'Airport Terminal',
         elevation: 10,
-        doors: [],
         images: [
           {
             data: airportPng,
@@ -189,8 +59,9 @@ export default async function buildingMap(): Promise<RomiCore.BuildingMap> {
             yaw: 0,
           },
         ],
+        doors: airportMap.doors,
+        places: airportMap.places,
         nav_graphs: [],
-        places: [],
         wall_graph: {
           name: 'wallgraph',
           vertices: [],
@@ -199,6 +70,6 @@ export default async function buildingMap(): Promise<RomiCore.BuildingMap> {
         },
       },
     ],
-    lifts: lifts
+    lifts: officeMap.lifts
   };
 }
